refactor: Merge without_note and without_since
